mzconduct: merge into mzcompose
Merge mzconduct and mzcompose. Where you previously ran

    $ bin/mzconduct chbench --workflow=ci

you now run:

    $ cd demo/chbench && ./mzcompose run ci

The hope is that the new API is less confusing to users. Everything
happens locally in the demo/test directory, so you no longer need to
reach out to ../../bin/mzconduct for some commands and ./mzcompose for
others.

The code is also greatly simplified, since we no longer have two
separate programs that need to parse and understand mzcompose.yml files.
