--- ci/plugins/mzconduct/hooks/pre-exit
@@ -1,25 +1,0 @@
-#!/bin/bash
-
-# Copyright Materialize, Inc. All rights reserved.
-#
-# Use of this software is governed by the Business Source License
-# included in the LICENSE file at the root of this repository.
-#
-# As of the Change Date specified in that file, in accordance with
-# the Business Source License, use of this software will be governed
-# by the Apache License, Version 2.0.
-
-set -euo pipefail
-
-echo "~~~ :docker: Cleaning up after mzconduct" >&2
-
-test_name="$BUILDKITE_PLUGIN_MZCONDUCT_TEST"
-compose_file="$(bin/ci-builder run stable bin/mzconduct show dir "$test_name")/mzcompose.yml"
-
-run() {
-    bin/ci-builder run stable bin/mzcompose --mz-quiet -f "$compose_file" "$@"
-}
-
-run kill
-run rm --force -v
-run down --volumes
