--- ci/plugins/mzconduct/plugin.yml
@@ -1,21 +1,0 @@
-# Copyright Materialize, Inc. All rights reserved.
-#
-# Use of this software is governed by the Business Source License
-# included in the LICENSE file at the root of this repository.
-#
-# As of the Change Date specified in that file, in accordance with
-# the Business Source License, use of this software will be governed
-# by the Apache License, Version 2.0.
-
-name: mzconduct
-description: Runs an mzconduct file
-author: "Materialize, Inc."
-requirements: []
-configuration:
-  properties:
-    test:
-      type: string
-    workflow:
-      type: string
-  required: ["test", "workflow"]
-  additionalProperties: false
