+++ demo/chbench/bin/ingest_bench
@@ -10,11 +10,6 @@
 
 set -euo pipefail
 
-MTRLZ_GIT_ROOT="${MTRLZ_GIT_ROOT:-$(git rev-parse --show-toplevel)}"
-if [[ "${PWD}" != "${MTRLZ_GIT_ROOT}" ]]; then
-    cd "${MTRLZ_GIT_ROOT}"
-fi
-
 chbench_seconds=43200
 cleanup="no"
 num_measurements=6
@@ -87,15 +82,15 @@ main() {
     export NUM_WAREHOUSES="${warehouses}"
     export CHBENCH_RUN_SECONDS="${chbench_seconds}"
 
-    ./bin/mzconduct run chbench -w setup-ingest-benchmark
+    ./mzcompose run setup-ingest-benchmark
 
     for i in $(seq 1 "${num_measurements}"); do
         echo "Conducting measurement ${i} of ${num_measurements}"
-        ./bin/mzconduct run chbench -w measure-ingest-performance
+        ./mzcompose run measure-ingest-performance
     done
 
     if [[ "${cleanup}" = "yes" ]]; then
-        ./bin/mzconduct down chbench
+        ./mzcompose down
     fi
 }
