+++ demo/chbench/bin/snapshot_bench
@@ -10,11 +10,6 @@
 
 set -euo pipefail
 
-MTRLZ_GIT_ROOT="${MTRLZ_GIT_ROOT:-$(git rev-parse --show-toplevel)}"
-if [[ "${PWD}" != "${MTRLZ_GIT_ROOT}" ]]; then
-    cd "${MTRLZ_GIT_ROOT}"
-fi
-
 cleanup="no"
 data_dir="${HOME}/chbench_snapshot"
 num_measurements=6
@@ -98,16 +93,16 @@ main() {
     export PEEKER_QUERIES="${peeker_queries}"
 
     if [[ "${setup}" = "yes" ]]; then
-        ./bin/mzconduct run chbench -w setup-replay-benchmark
+        ./mzcompose run setup-replay-benchmark
     fi
 
     for i in $(seq 1 "${num_measurements}"); do
         echo "Conducting measurement ${i} of ${num_measurements}"
-        ./bin/mzconduct run chbench -w measure-ingest-performance
+        ./mzcompose run measure-ingest-performance
     done
 
     if [[ "${cleanup}" = "yes" ]]; then
-        ./bin/mzconduct down chbench
+        ./mzcompose down chbench
     fi
 }
