+++ misc/python/materialize/ui.py
@@ -99,7 +99,7 @@ def timeout_loop(timeout: int, tick: int = 1) -> Generator[float, None, None]:
 def log_in_automation(msg: str) -> None:
     """Log to a file, if we're running in automation"""
     if env_is_truthy("MZ_IN_AUTOMATION"):
-        with open("/tmp/mzconduct.log", "a") as fh:
+        with open("/tmp/mzcompose.log", "a") as fh:
             now = datetime.datetime.now().isoformat()
             print(f"[{now}] {msg}", file=fh)
